# docs/swagger.yaml
definitions:
...
  handler.Item:
    properties:
      authors:
        items:
          $ref: '#/definitions/handler.Person'
        type: array
      categories:
        items:
          type: string
        type: array
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      enclosures: {}
      feed_id:
        type: string
      guid:
        type: string
      id:
        type: string
      image: {}
      liked:
        type: boolean
      liked_at:
        type: string
      link:
        type: string
      links:
        items:
          type: string
        type: array
      published_parsed:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_parsed:
        type: string
    type: object
...
  handler.ListItemsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.Item'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total_count:
        type: integer
    type: object
  handler.Person:
    properties:
      email:
        description: 'example: jane@example.com'
        type: string
      name:
        description: 'example: Jane Doe'
        type: string
    type: object
info:
  contact: {}
  description: Swagger API documentation for Gazette.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gazette API
  version: 0.1.0
paths:
...
  /api/items/:
    get:
      description: Retrieves items liked by the user, paginated.
      parameters:
      - description: Max number of items
        in: query
        name: limit
        required: true
        type: integer
      - description: Number of items to skip
        in: query
        name: offset
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListItemsResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List liked items
      tags:
      - Items
  /api/items/{itemID}:
    get:
      description: Retrieves an item by ID, including like status for the user.
      parameters:
      - description: Item UUID
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Item'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get item
      tags:
      - Items
  /api/items/{itemID}/like:
    delete:
      description: Deletes the like record for the current user on an item.
      parameters:
      - description: Item UUID
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Unlike item
      tags:
      - Items
  /items/{itemID}/like:
    post:
      description: Creates a like record for the current user on an item.
      parameters:
      - description: Item UUID
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "200":
          description: liked_at
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Like item
      tags:
      - Items
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"